yassir:pay process implementation

+++ src/main/java/com/fstg/bookerpaymentservice/application/dto/PaymentDto.java
@@ -1,13 +1,14 @@
 package com.fstg.bookerpaymentservice.application.dto;
 
+import com.fstg.bookerpaymentservice.domain.payment.PaymentCreateProcessInput;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-
 import org.springframework.beans.BeanUtils;
 
-import com.fstg.bookerpaymentservice.domain.payment.PaymentCreateProcessInput;
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,10 +17,8 @@ public class PaymentDto {
 
     private Long id;
     private String reference;
-    private double amount;
+    private BigDecimal amount;
     private String orderReference;
-
-
 
     public static PaymentCreateProcessInput to(PaymentDto paymentDto) {
         PaymentCreateProcessInput paymentCreateProcessInput = new PaymentCreateProcessInput();
